Separate packages imported by installer from packages importing controller-runtime client
Right now, the installer imports the following packages:

```go
ccaws "github.com/openshift/cloud-credential-operator/pkg/aws"
credvalidator "github.com/openshift/cloud-credential-operator/pkg/controller/utils"
```

Disadvantage of the current layout in cloud-controller-operator repo causes
sigs.k8s.io/controller-runtime/pkg/client package to be imported as well.
The newer version of sigs.k8s.io/controller-runtime/pkg/client package requires new
data type(s) from k8s.io/apimachinery which are not available in k8s 1.12 (which
is currently imported by the installer).

Also most of the exported methods/functions were used solely by aws part
of the cloud-controller-operator. Moving aws specific bits under pkg/aws
improves importability by other projects and removes the need
for importing sigs.k8s.io/controller-runtime/pkg/client bits that are
not used be the installer anyway.
