+++ pkg/aws/client.go
@@ -17,14 +17,6 @@ limitations under the License.
 package aws
 
 import (
-	"context"
-	"fmt"
-
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/types"
-
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
 	awssdk "github.com/aws/aws-sdk-go/aws"
 	"github.com/aws/aws-sdk-go/aws/credentials"
 	"github.com/aws/aws-sdk-go/aws/request"
@@ -32,11 +24,6 @@ import (
 	"github.com/aws/aws-sdk-go/service/iam"
 	"github.com/aws/aws-sdk-go/service/iam/iamiface"
 	"github.com/openshift/cloud-credential-operator/version"
-)
-
-const (
-	awsCredsSecretIDKey     = "aws_access_key_id"
-	awsCredsSecretAccessKey = "aws_secret_access_key"
 )
 
 //go:generate mockgen -source=./client.go -destination=./mock/client_generated.go -package=mock
@@ -128,29 +115,4 @@ func NewClient(accessKeyID, secretAccessKey []byte, infraName string) (Client, e
 	return &awsClient{
 		iamClient: iam.New(s),
 	}, nil
-}
-
-func LoadCredsFromSecret(kubeClient client.Client, namespace, secretName string) ([]byte, []byte, error) {
-
-	secret := &corev1.Secret{}
-	err := kubeClient.Get(context.TODO(),
-		types.NamespacedName{
-			Name:      secretName,
-			Namespace: namespace,
-		},
-		secret)
-	if err != nil {
-		return nil, nil, err
-	}
-	accessKeyID, ok := secret.Data[awsCredsSecretIDKey]
-	if !ok {
-		return nil, nil, fmt.Errorf("AWS credentials secret %v did not contain key %v",
-			secretName, awsCredsSecretIDKey)
-	}
-	secretAccessKey, ok := secret.Data[awsCredsSecretAccessKey]
-	if !ok {
-		return nil, nil, fmt.Errorf("AWS credentials secret %v did not contain key %v",
-			secretName, awsCredsSecretAccessKey)
-	}
-	return accessKeyID, secretAccessKey, nil
 }
